demos/linux/devicetrees/imx8mn-ddr3l-evk/parseDT/requirements.txt
# fdt==0.3.3
pyfdt==0.3
pydevicetree==0.0.13
